Extracted InterpreterSetting from InterpreterFactory
